The BluKane mobile application is an obstacle avoidance mobile application that is used to help visually impaired individuals navigate to their destinations.
The BluKane is a low-cost navigation tool for the visually impaired.  It works by the reading data sent from an Ultrasonic Distance Sensor, via a bluetooth module (HC-06). The wired sensor and bluetooth module, should be mounted on the user's traditional white cane and the mobile application, should be used to connect to it.